11/2 progress report
- modules imported from old rewrite
- general cleanup
- stored more useful information about users in punishments etc

+++ Administrator.Core/Extensions/DateTimeExtensions.cs
@@ -24,6 +24,6 @@ public static class DateTimeExtensions
         IanaTimeZoneMap = dict;
     }
 
-    public static DateTimeOffset AddWeeks(this DateTimeOffset dto, int weeks)
+    public static DateTimeOffset AddWeeks(this DateTimeOffset dto, double weeks)
         => dto.AddDays(weeks * 7);
 }
